test/CMakeLists.txt: set RPATH via the linker.
I was having a problem when deploying the test executable together with the libFreeImage.so to a docker folder. The shared library couldn't be found. The change going in this commit adds the test executable directory to the RPATH (the list of paths where shared libraries are found) so that libFreeImage.so is now found.